3.2.1, 3.2.2 done 3.2.3, 3.2.4 WIP
