+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 Project.class
 ojdbc7o.jar
+*.class
