--- .idea/hotswap_agent.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="HotSwapAgentPluginSettingsProvider">
-    <option name="agentPath" value="C:\Users\Owner\AppData\Roaming\JetBrains\IdeaIC2023.2\plugins\hotswap-agent-intellij-plugin\lib\agent\hotswap-agent-1.4.1.jar" />
-  </component>
-</project>
