+++ README.md
@@ -1,7 +1,33 @@
 # CSCI3170 project
 
 ### how to run
+#### TRY THIS FIRST
+Windows:
 ```
-javac Project.java
+Ctrl + Shift + B
+```
+
+MacOS:
+```
+⌘ + ⇧ + b
+```
+
+Windows:
+```
+javac -d ./out/ -classpath ".;ojdbc7.jar" "./csci3170project/*.java"
+java -classpath ".;ojdbc7.jar" csci3170project.Project
+```
+
+Linux/Mac:
+```
+javac -d ./out/ -classpath ".:ojdbc7.jar" "./csci3170project/*.java"
+java -classpath ".:ojdbc7.jar" csci3170project.Project
+```
+
+
+Old stuff
+```
+javac Project.java 
 java -classpath ./ojdbc7.jar:./ Project
 ```
+
